% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.r
\name{fit.BRNN}
\alias{fit.BRNN}
\title{Modeling and predicting using Bayesian Regularization Neural Networks(BRNN)}
\usage{
fit.BRNN(
  trainMarkerMat,
  trainPheVal,
  predictMarkerMat,
  NAImpute = F,
  outputModel = FALSE,
  verbose = TRUE,
  neurons = 4,
  epochs = 30,
  cpus = 1,
  ...
)
}
\description{
Modeling and predicting using Bayesian Regularization Neural Networks(BRNN)
}
