% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{fit.EM}
\alias{fit.EM}
\title{Modeling and predicting using methods EM model.}
\usage{
fit.EM(trainedMarkerMat, trainedPheVal, predictMarkerMat,
  effectConcider = "A", outputModel = F, Z = NULL, X = NULL,
  iters = 20, draw = FALSE, silent = TRUE, constraint = TRUE,
  init = NULL, forced = NULL, tolpar = 1e-04, tolparinv = 1e-06, ...)
}
\arguments{
\item{trainedMarkerMat}{(numeric)a matrix, each row is the each training sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{trainedPheVal}{(numeric)the phenotype Value of each individual.}

\item{predictMarkerMat}{(numeric)a matrix, each row is the each testing sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{effectConcider}{(string)if Z = NULL, random effects are auto generated.}

\item{outputModel}{(logical)if true, return the list of training model.}

\item{Z}{A 2-level list,incidence matrices and var-cov matrices for random effects. This works for ONE OR MORE random effects. This needs to be provided as a 2-level list structure, defult NULL.}

\item{X}{(numeric, matrix) design matrix related to the parameters not to be shrunk (i.e. fixed effects in the mixed model framework),defult no shrink.}

\item{iters}{(numeric)a scalar value indicating how many iterations have to be performed if the optimization algorithms. There is no rule of tumb for the number of iterations but less than 8 is usually enough, default 20.}

\item{draw}{(logical)indicating if a plot of updated values for the variance components and the likelihood should be drawn or not. The default is TRUE. COMPUTATION TIME IS SMALLER IF YOU DON'T PLOT SETTING draw=FALSE, default FALSE.}

\item{silent}{(logical)indicating if the function should draw the progress bar or iterations performed while working or should not be displayed, default FALSE}

\item{constraint}{(logical)indicating if the program should use the boundary constraint when one or more variance component is close to the zero boundary. The default is TRUE but needs to be used carefully. It works ideally when few variance components are close to the boundary but when there are too many variance components close to zero we highly recommend setting this parameter to FALSE since is more likely to get the right value of the variance components in this way, default FALSE.}

\item{init}{vector of initial values for the variance components. By default this is NULL and variance components are estimated by the method selected, but in case the user want to provide initial values this argument is functional.}

\item{forced}{a vector of numeric values for variance components including error if the user wants to force the values of the variance components. On the meantime only works for forcing all of them and not a subset of them. The default is NULL, meaning that variance components will be estimated by REML/ML.}

\item{tolpar}{tolerance parameter for convergence in the models.}

\item{tolparinv}{tolerance parameter for matrix inversion in the models.}

\item{...}{other parameters, details see package sommer.}
}
\value{
a list including model and prediction result (outputModel = TRUE)
a array of prediction result (outputModel = FALSE)
}
\description{
This function can fit EM model and export the prediction values of testing sets.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## use RR model to modeling and predict ##
EM_Res <- fit.EM(trainedMarkerMat = Markers[1:200,], trainedPheVal = phenotype, 
                   predictMarkerMat = Markers[201:242,], method = "EM", 
                   effectConcider = "A", iters = 20,
                   )
}
}
\seealso{
\pkg{sommer}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{AI,}
\keyword{EM,}
\keyword{EMMA,}
\keyword{NR,}
\keyword{model}
\keyword{predict}
