% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{fit.mmer}
\alias{fit.mmer}
\title{Modeling and predicting using methods AI, NR , EM , EMMA.}
\usage{
fit.mmer(trainedMarkerMat, trainedPheVal, predictMarkerMat, Z = NULL,
  X = NULL, method = "NR", effectConcider = "A", outputModel = FALSE,
  iters = 20)
}
\arguments{
\item{trainedMarkerMat}{(numeric)a matrix, each row is the each training sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{trainedPheVal}{(numeric)the phenotype Value of each individual.}

\item{predictMarkerMat}{(numeric)a matrix, each row is the each testing sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{Z}{A 2-level list,incidence matrices and var-cov matrices for random effects. This works for ONE OR MORE random effects. This needs to be provided as a 2-level list structure, defult NULL.}

\item{X}{(numeric, matrix) design matrix related to the parameters not to be shrunk (i.e. fixed effects in the mixed model framework),defult no shrink.}

\item{method}{(string)select the algorithm to fit the model including NR, AI, EM, EMMA."AI and EM method has been discontinued because of its instability. Try 'EMMA' and 'NR'.See details in the G2P help page.If you insist AI & EM, please use function fit.EM and fit.AI to debug.}

\item{effectConcider}{(string)if Z = NULL, random effects are auto generated.}

\item{outputModel}{(logical)if true, return the list of training model.}

\item{iters}{(numeric)a scalar value indicating how many iterations have to be performed if the optimization algorithms. There is no rule of tumb for the number of iterations but less than 8 is usually enough, default 20.}

\item{cpus}{(numeric)number of cores to use when the user passes multiple responses in the model for a faster execution of univariate models. The default is 1}

\item{verbose}{(logical)if TRUE, the iteration history is printed, default FALSE.}

\item{rmExistingFiles}{(logical)if TRUE removes existing output files from previous runs, default TRUE.}

\item{...}{other parameters, details see package sommer.}
}
\value{
a list including model and prediction result (outputModel = TRUE)
a array of prediction result (outputModel = FALSE)
}
\description{
This function can fit AI, NR , EM , EMMA model and export the prediction values of testing sets.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## use RR model to modeling and predict ##
EM_Res <- fit.mmer(trainedMarkerMat = Markers[1:200,], trainedPheVal = phenotype, 
                   predictMarkerMat = Markers[201:242,], method = "EM", 
                   effectConcider = "A", iters = 20,
                   )
}
}
\seealso{
\pkg{sommer}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{AI,}
\keyword{EM,}
\keyword{EMMA,}
\keyword{NR,}
\keyword{model}
\keyword{predict}
