% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{fit.RKHS}
\alias{fit.RKHS}
\title{Modeling and predicting using Reproducing Kernel Hilbert Space(RKHS).}
\usage{
fit.RKHS(trainedMarkerMat, trainedPheVal, predictMarkerMat,
  outputModel = FALSE, nIter = 1500, burnIn = 500, thin = 5,
  saveAt = "", S0 = NULL, df0 = 5, R2 = 0.5, weights = NULL,
  verbose = FALSE, rmExistingFiles = TRUE, groups = NULL)
}
\arguments{
\item{trainedMarkerMat}{(numeric)a matrix, each row is the each training sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{trainedPheVal}{(numeric)the phenotype Value of each individual.}

\item{predictMarkerMat}{(numeric)a matrix, each row is the each testing sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{outputModel}{(logical)if true, return the list of training model.}

\item{nIter, burnIn, thin}{(integer)the number of iterations, burn-in and thinning,default nIter 1500, burnIn 500, thin 5.}

\item{saveAt}{(string)this may include a path and a pre-fix that will be added to the name of the files that are saved as the program runs,default "".}

\item{S0, df0}{(numeric)the scale parameter for the scaled inverse-chi squared prior assigned to the residual variance, only used with Gaussian outcomes. In the parameterization of the scaled-inverse chi square in BGLR the expected values is S0/(df0-2). The default value for the df parameter is 5. If the scale is not specified a value is calculated so that the prior mode of the residual variance equals var(y)*R2 (see below). For further details see the vignettes in the package or http://genomics.cimmyt.org/BGLR-extdoc.pdf. Default S0 NULL,df0 5.}

\item{R2}{(numeric, (0,1))the proportion of variance that one expects, a priori, to be explained by the regression. Only used if the hyper-parameters are not specified; if that is the case, internaly, hyper-paramters are set so that the prior modes are consistent with the variance partition specified by R2 and the prior distribution is relatively flat at the mode. For further details see the vignettes in the package or http://genomics.cimmyt.org/BGLR-extdoc.pdf. Defult 0.5}

\item{weights}{(numeric, n)a vector of weights, may be NULL. If weights is not NULL, the residual variance of each data-point is set to be proportional to the square of the weight. Only used with Gaussian outcomes.}

\item{verbose}{(logical)if TRUE, the iteration history is printed, default FALSE.}

\item{rmExistingFiles}{(logical)if TRUE, removes existing output files from previous runs, default TRUE.}

\item{groups}{(factor)a vector of the same length of y that associates observations with groups, each group will have an associated variance component for the error term.}
}
\value{
a list including model and prediction result (outputModel = TRUE)
a array of prediction result (outputModel = FALSE)
}
\description{
This function can fit RKHS model and export the prediction value of testing sets.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## use RR model to modeling and predict ##
RKHS_Res <- fit.RKHS(trainedMarkerMat = Markers, trainedPheVal = phenotype, 
                     predictMarkerMat = Markers[1:10,],nIter = 1500, burnIn = 500)
}
}
\seealso{
\pkg{BGLR}
}
\author{
Chuang Ma  , Qian Cheng , Zhixu Qiu
}
\keyword{RKHS,}
\keyword{model}
\keyword{predict}
