% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{fit.RR}
\alias{fit.RR}
\title{Modeling and Predicting using ridge regression}
\usage{
fit.RR(trainedMarkerMat, trainedPheVal, predictMarkerMat, cpus = 1,
  outputModel = FALSE)
}
\arguments{
\item{trainedMarkerMat}{(numeric,matrix)each row is the each training sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{trainedPheVal}{(numeric)the phenotype Value of each individual.}

\item{predictMarkerMat}{(numeric,matrix)each row is the each testing sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{cpus}{(integer)number of cpu cores to use for calculations (only available in UNIX-like operating systems), default 1.}

\item{outputModel}{(logical)if true, return the list of training model.}
}
\value{
a list including model and prediction result (outputModel = TRUE)
a array of prediction result (outputModel = FALSE)
}
\description{
This function can fit ridge regression model and export the prediction value of testing sets.
}
\examples{
\dontrun{
## Not run: 
## Load example data ##
data(GYSS)

## use RR model to modeling and predict ##
rr_Res <- fit.RR(trainedMarkerMat = Markers, trainedPheVal = phenotype,
                 predictMarkerMat = Markers[1:10,], cpus = 1 )
}
}
\seealso{
\pkg{rrBLUP}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{RR}
\keyword{model}
\keyword{predict}
