% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMerge.r
\name{GSIntegrate}
\alias{GSIntegrate}
\title{Prediction Integration from Two Methods}
\usage{
GSIntegrate(predResMat, ratio, autoOptimize = F)
}
\arguments{
\item{predResMat}{(numeric, matrix)the prediction results of algorithms which you want to merge, the first column is the real value of trait.}

\item{ratio}{(numeric,array) the weights of every algorithms.}

\item{autoOptimize}{(logical)if auto select two method results from multi-results and then compute the mean of two methods results (1:1), default FALSE.}
}
\value{
a matrix:  
involve real value of trait, merge algorithms and the merge result.
}
\description{
This function provides a strategy to integrate the results of two or more algorithms.
}
\details{
The predResMat must including real value in first clumn, and if you set "autoOptimize = T", the count of algorithms must more than 2.

In this function, if autoOptimize = T, the final two algorithms merge are selected from multi methods by following strategy:
Firstly, compute the pearson's correlation of predResMat, choose the best correlation between real value and prediction scores, named method 1.
Secondly, choose the best correlation between method 1 and other methods, named method 2.
Finally, merge method 1 and method 2 with 1:1(mean).

This function auto merge only provide integrate base \bold{pearson's correlation} evaluation.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## cross validation ##
predlist <- G2PCrossvalidation(cross = 10,seed = 1 , cpus = 3, markers = Markers,
                pheVal = phenotype, modelMethods = c("BayesA","BayesB","BayesC","rrBLUP", "RFR"),
                outputModel = FALSE)
resultMat <- resultsMerge(predlist)

## merge ##
inter <- GSIntegrate(predResMat = resultMat[,1:6],
                     ratio = c(2,3,4,4,5), autoOptimize = F)
interAuto <- GSIntegrate(predResMat = resultMat, autoOptimize = T,
                         allMethodPredResMat = resultMat)
}
}
\seealso{
\link{GSEnsemble}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu, Jie Song
}
\keyword{integration,}
\keyword{merge}
