% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{GSReModel}
\alias{GSReModel}
\title{Fit Regression Model}
\usage{
GSReModel(markers, pheVal, modelMethods, K = 8, eta = 0.7,
  select = "pls2", fit = "simpls", scale.x = FALSE, scale.y = FALSE,
  eps = 1e-04, trace = FALSE, maxstep = 100, alpha = 1, X = NULL,
  family = gaussian(link = identity), lambda = NULL, tol.err = 1e-06,
  tol.conv = 1e-08, weights = NULL, ...)
}
\arguments{
\item{markers}{(numeric)a matrix, each row is the each individual's SNP genotypes informations.Genotypes should be coded as {0,1,2}or{-1,0,1};0(-1) represent AA(homozygote),2(1) represent BB(homozygote) and 1(0) represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{pheVal}{(numeric)the phenotype value of each individual.}

\item{K}{(integer)SPLS model parameter: number of hidden components.}

\item{eta}{(numeric)SPLS model parameter: thresholding parameter. eta should be between 0 and 1.}

\item{select}{(character)SPLS model parameter: PLS algorithm for variable selection. Alternatives are "pls2" or "simpls". Default is "pls2".}

\item{fit}{(character)SPLS model parameter: PLS algorithm for model fitting. Alternatives are "kernelpls", "widekernelpls", "simpls", or "oscorespls". Default is "simpls".}

\item{scale.x}{(character)SPLS model parameter: scale predictors by dividing each predictor variable by its sample standard deviation?}

\item{scale.y}{(character)SPLS model parameter: scale responses by dividing each response variable by its sample standard deviation?}

\item{eps}{(character)SPLS model parameter: an effective zero. Default is 1e-4.}

\item{trace}{(logical)SPLS model parameter: print out the progress of variable selection?}

\item{maxstep}{(integer)SPLS model parameter: maximum number of iterations when fitting direction vectors. Default is 100.}

\item{alpha}{(numeric)LASSO model parameter: the elasticnet mixing parameter.Detail in glmnet.}

\item{X}{(numeric, matrix) design matrix related to the parameters not to be shrunk (i.e. fixed effects in the mixed model framework),defult no shrink.}

\item{family}{the distribution family of y, see help('family') for more details.}

\item{lambda}{the shrinkage parameter determines the amount of shrinkage. Default is NULL meaning that it is to be estimated along with other model parameters.}

\item{tol.err}{internal tolerance level for extremely small values; default value is 1e-6.}

\item{tol.conv}{tolerance level in convergence; default value is 1e-8.}

\item{...}{arguments passed to or from other package.}
}
\value{
A regression model which is enable to predict.
}
\description{
This function can fit several regression models of genomic selection including SPLS, rrBLUP, LASSO and bigRR.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## Fit rrBLUP model ##
rrBLUP_model <- GSReModel(markers = Markers,pheVal = phenotype,modelMethods = "rrBLUP")
}
}
\seealso{
\pkg{rrBLUP}
\pkg{hglm}
\pkg{glmnet}
\pkg{spls}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{LASSO,}
\keyword{SPLS,}
\keyword{bigRR}
\keyword{model,}
\keyword{regression}
\keyword{rrBLUP,}
