% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\docType{data}
\name{GYSS}
\alias{GYSS}
\title{Example Data for G2P}
\description{
The data of maize yield under drought stressed.(SNP genotypes informations) 
 \itemize{
 \item{Markers} {A numeric matrix, each row is the each individual's SNP genotypes informations.}
 \item{phenotype} {The real phenotype Value of each individual.}
 }
}
\examples{
\dontrun{
## load maize yield data sets
data(GYSS)
}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{datasets}
