% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\docType{data}
\name{Markers}
\alias{Markers}
\title{Example Data for G2P}
\format{matrix}
\description{
A numeric matrix.
 \itemize{
 \item{each row represents each sample}
 \item{each column represents each SNP locus}
 }
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{datasets}
