% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{Example Data for G2P}
\format{vector}
\description{
The real phenotype Value of each individual
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{,}
\keyword{data}
\keyword{phenotype}
\keyword{value}
