% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{randomSeed}
\alias{randomSeed}
\title{Generate Random Seed}
\usage{
randomSeed()
}
\value{
(numeric) A random seed
}
\description{
This funcation is appplied for generating random seed with current system time
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
