% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{evaluateGS}
\alias{evaluateGS}
\title{evaluateGS}
\usage{
evaluateGS(realScores, predScores, Probability = FALSE, evalMethod = "RE",
  Beta = 1, BestIndividuals = "top", topAlpha = 1:90, allNDCG = F,
  globalAlpha = F, probIndex = NULL)
}
\arguments{
\item{realScores}{(numeric,vector)vector is the real breeding values of the validation individual for a trait.}

\item{predScores}{(numeric,vector or matrix)the prediction breeding value predicted by genomic selection model of the individuals.}

\item{Probability}{(logical)whether the predScores is probability? Default FALSE.}

\item{evalMethod}{(character)the evaluation methods selected to evaluate, which include "pearson", "kendall", "spearman", "MSE","R2"
"RE", "Kappa", "AUC","AUCpr","accuracy","F1","meanNDCG", "NDCG".}

\item{Beta}{(numeric)the parameter of "F1".}

\item{BestIndividuals}{(character)the position of expected phenotype in whole phenotypic data set."top","buttom" or "middle",default "top".}

\item{topAlpha}{(numeric,(0,100])a vector is the proportion of excellent individuals,default 1:90.}

\item{globalAlpha}{(logical)indicates if evaluate global methods(pearson, kendall, spearman, MSE and R2) by alpha,default FALSE.}

\item{probIndex}{(integer)indicates the column index which prediction result is probability.}
}
\value{
a list inculding  evaluation results with methods which user selected.
}
\description{
this function is used to evaluete the accuracy of predicted by genomic selection model.
}
\examples{
\dontrun{
data(GYSS)
########## predicting breeding value
predlist <-  G2PCrossValidation(cross = 10,seed = 1 ,cpus = 3,markers  = Markers,pheVal  = phenotype,
                 modelMethods = c("rrBLUP","RFC"),outputModel = FALSE)
predMartix <- NULL
for(ii in 1:10){predMartix <- rbind(predMartix,predlist[[ii]])}
######## evaluate the accuracy of the prediction result

evaluareTest <- evaluateGS(realScores = predMartix[,1], predScores = predMartix[,2:3], 
                           evalMethod = c("pearson", "kendall","spearman","RE","Kappa",
                                         "AUC","AUCpr","NDCG","meanNDCG",
                                         "MSE","R2","F1","accuracy"),topAlpha = 1:90, probIndex = 2)
}
}
\author{
Chuang Ma, Zhixu Qiu, Qian Cheng
}
\keyword{AUC,}
\keyword{AUCpr,}
\keyword{F1,}
\keyword{Kappa,}
\keyword{MSE,}
\keyword{NDCG}
\keyword{R2,}
\keyword{RE,}
\keyword{accuracy,}
\keyword{evaluation,}
\keyword{kendall,}
\keyword{meanNDCG,}
\keyword{pearson,}
\keyword{spearman,}
