% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{feature_assess}
\alias{feature_assess}
\title{Feature Selection}
\usage{
feature_assess(markers, phenotype, method = c("rrBLUP", "Gini", "Accuracy"),
  ntree = 500, importance = TRUE, posPercentage = 0.4,
  BestIndividuals = c("top", "middle", "buttom"))
}
\arguments{
\item{markers}{(numeric, matrix)row is sample well column is SNP information (feature).Genotypes should be coded as {0,1,2};0 represent AA(homozygote),2 represent BB(homozygote) and 1 represent AB(heterozygote);missing (NA) alleles are not allowed.}

\item{phenotype}{(numeric)the phenotype value of each individual.}

\item{method}{(character)the method of feature selction including "Gini" "Accuracy" "rrBLUP", default "RR-BLUP"}

\item{ntree}{(numeric)the number of random forest decision tree, default 500}

\item{importance}{(logical)whether the results of variable importance,default TRUE}

\item{posPercentage}{(numeric,[0,1])phenotype of extreme individuals which expected, default 0.4}

\item{BestIndividuals}{(character)the position of expected phenotype in whole phenotypic data set."top","buttom" or "middle",defult "top".}
}
\value{
A numeric mode score of each position of SNPs
}
\description{
This function scores each marker,so that reduce the data dimension and perform feature selection.(Methods including Gini,Accuracy and rrBLUP).
}
\examples{
\dontrun{
data(GYSS)
## feature selection with Gini ##
Gini_selection <- feature_assess(markers = Markers, phenotype = phenotype, method = "Gini", 
                                 ntree = 500, importance = TRUE, posPercentage = 0.40,
                                 BestIndividuals = "top")

## feature selection with Acc ##
Acc_selection <- feature_assess(markers = Markers, phenotype = phenotype, method = "Accuracy", 
                                ntree = 500, importance = TRUE, posPercentage = 0.40,
                                BestIndividuals = "top")

## feature selection with rrBLUP ##
rrBLUP_selection <- feature_assess(markers = Markers, phenotype = phenotype, method = "rrBLUP", 
                                   posPercentage = 0.40, BestIndividuals = "top")
}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{Accuracy,}
\keyword{Gini,}
\keyword{RR-BLUP}
\keyword{feature}
\keyword{selction,}
