% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{fit.BRNN}
\alias{fit.BRNN}
\title{Modeling and predicting using Bayesian Regularization Neural Networks(BRNN)}
\usage{
fit.BRNN(trainedMarkerMat, trainedPheVal, predictMarkerMat,
  outputModel = FALSE, verbose = TRUE, neurons = 4, epochs = 30,
  cpus = 1, ...)
}
\arguments{
\item{trainedMarkerMat}{(numeric)a matrix, each row is the each training sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote),2 represent BB(homozygote) and 1 represent AB(heterozygote);missing (NA) alleles are not allowed.}

\item{trainedPheVal}{(numeric)the phenotype Value of each individual.}

\item{predictMarkerMat}{(numeric)a matrix, each row is the each testing sets individual's SNP genotypes informations. Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote),2 represent BB(homozygote) and 1 represent AB(heterozygote);missing (NA) alleles are not allowed.}

\item{outputModel}{(logical)if true, return the list of training model.}

\item{verbose}{(logical)if TRUE, will print iteration history.}

\item{neurons}{(integer)indicates the number of neurons,defult 4.}

\item{epochs}{(integer)maximum number of epochs(iterations) to train, default 30.}

\item{cpus}{(integer)number of cpu cores to be used for calculations (only available in UNIX-like operating systems), default 1.}

\item{...}{other parameters, details see package brnn}
}
\value{
a list including model and prediction result (outputModel = TRUE)
a array of prediction result (outputModel = FALSE)
}
\description{
This function can fit BRNN model and export the prediction value of testing sets.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## use RR model to modeling and predict ##
BRNN_Res <- fit.BRNN(trainedMarkerMat = Markers,trainedPheVal = phenotype,
                     predictMarkerMat = Markers[1:10,],cpus = 1 )
}
}
\seealso{
\pkg{brnn}
}
\author{
Chuang Ma  , Qian Cheng , Zhixu Qiu
}
\keyword{BRNN,}
\keyword{model}
\keyword{predict}
