% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{GSDataQC}
\alias{GSDataQC}
\title{Genomic Selection Data Quality control}
\usage{
GSDataQC(markers, phenotype, impute = F, filter = F, NABound = 0.8,
  MAFBound = 0.005, imputeMethod = "mean", round = 4, k = 10,
  maxp = 1500, rowmax = 0.5, colmax = 0.8, rng.seed = 362436069)
}
\arguments{
\item{markers}{(numeric, matrix)row is sample well column is SNP information (feature).Genotypes should be coded as {0,1,2};0 represent AA(homozygote),2 represent BB(homozygote) and 1 represent AB(heterozygote);missing (NA) alleles are not allowed.}

\item{phenotype}{(numeric)the phenotype value of each individual.}

\item{impute}{(logical)if TRUE, imputation, default F.}

\item{filter}{(logical)if TRUE, filter the markers data with the setting of MAF and missing value.}

\item{NABound}{(numeric, [0,1])max missing value percentage.}

\item{MAFBound}{(numerix, [0,1])min MAF percentage in each marker local sit.}

\item{imputeMethod}{(character)the method of imputation, "mean", "median" or "KNN", default "mean".}

\item{round}{(numeric)rounds the values in its first argument to the specified number of decimal places, default 4.}

\item{k, maxp, rowmax, colmax, rng.seed}{(numeric)KNN method parameters.}
}
\value{
A list of the data quality information.
}
\description{
This function can examine and summary the quality of GS data. And can be used for imputation.
}
\examples{
\dontrun{
data(GYSS)
## generate missing value
misIndex <- sample(1:242000,100000)
Markers[misIndex] <- NA

## GSDataQC, not impute ##
QCRes <- GSDataQC(markers = Markers, phenotype = phenotype, impute = F)

## GSDataQC, not impute ##
QCResImpute <- GSDataQC(markers = Markers, phenotype = phenotype, impute = T, 
                                    imputeMethod = "mean")
}
}
\seealso{
\pkg{impute}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{Accuracy,}
\keyword{Gini,}
\keyword{RR-BLUP}
\keyword{feature}
\keyword{selction,}
