% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEnsemble.R
\name{GSEnsemble}
\alias{GSEnsemble}
\title{Methods Ensemble From Two or More Methods}
\usage{
GSEnsemble(predMat, nrandom = 10, evalMethods, by = 0.1, evaluation = T,
  topAlpha = 15, ...)
}
\arguments{
\item{predMat}{(numeric, matrix)the prediction results of algorithms which you want to merge, the first column is the real value of trait.}

\item{nrandom}{(integer)the repeat number of stacking, default 10.}

\item{evalMethods}{(character)ensemble base which evaluation methods.}

\item{by}{(numeric,(0,1))the radio window of ensemble, the smaller "by", the higher accuracy of ensemble. Default 0.1.}

\item{evaluation}{(logical)if evaluate finalMat with evalMethods, default TRUE.}

\item{topAlpha}{(numeric)the parameter of threshold evaluation methods, see also function evaluateGS. In this function, indicates the best ensemble base threshold when evalMethods is threshold methods.}

\item{...}{arguments passed to or from other functions.}
}
\value{
a list:  
$BestWeight The best weight of methods in all repeat
$finalMat The final matrix cbind predMat with final ensemble score
$evalRes The evaluation results of finalMat with evalMethods
$weightMat A weight matrix including all repeats
$evalMat A evaluation results matrix including all repeats
}
\description{
This function provides a strategy to ensemble the results of two or more algorithms. It is a extension of GSIntegrate
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## cross validation ##
predlist <- G2PCrossvalidation(cross = 10,seed = 1 , cpus = 3, markers = Markers,
                pheVal = phenotype, modelMethods = c("BayesA","BayesB","BayesC","rrBLUP", "RFR"),
                outputModel = FALSE)
resultMat <- resultsMerge(predlist)

## merge ##
ensembleRes <- GSEnsemble(predMat = resultMat, nrandom = 10, evalMethods = "RE",
                          by = 0.1, evaluation = T, topAlpha = 15 )
}
}
\seealso{
\link{GSIntegrate}
\link{evaluateGS}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu, Jie Song
}
\keyword{integration,}
\keyword{merge}
