% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{GSmachine}
\alias{GSmachine}
\title{Fit machine learning model}
\usage{
GSmachine(markers, pheVal, modelMethods = "SVC", posPercentage = 0.4,
  BestIndividuals = c("top"), ntree = 500, nodesize = NULL,
  kernel = c("linear"), gamma = 1, cost = 2^(-9), ...)
}
\arguments{
\item{markers}{(numeric)a matrix, each row is the each individual's SNP genotypes informations.Genotypes should be coded as {0,1,2};0 represent AA(homozygote),2 represent BB(homozygote) and 1 represent AB(heterozygote);missing (NA) alleles are not allowed.}

\item{pheVal}{(numeric)the phenotype value of each individual.}

\item{modelMethods}{(character)alternative machine learning models. "SVR" and "SVC" from SVM, "RFR" and "RFC" from RF.}

\item{posPercentage}{(numeric,[0,1])phenotype of extreme individuals which expected, default 0.4.}

\item{BestIndividuals}{(character)the position of expected phenotype in whole phenotypic data set."top","buttom" or "middle",default "top".}

\item{ntree}{(integer)ramdomforest parameter. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times,default 500.}

\item{nodesize}{(integer)ramdomforest parameter Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time). Note that the default values are different for classification (1) and regression (5).}

\item{kernel}{(numeric)svm parameter the kernel used in training and predicting. You might consider changing some of the following parameters, depending on the kernel type.(linear,polynomial,sigmoid,radial)Default "linear".}

\item{gamma}{(numeric)svm parameter parameter needed for all kernels except linear, default 1.}

\item{cost}{(numeric)svm parameter cost of constraints violation, default: 2^(-9), it is the 'C'-constant of the regularization term in the Lagrange formulation.}

\item{...}{other parameters.}
}
\value{
a machine learning model
}
\description{
This function can fit several machine learning models of genomic selection such as "SVR", "SVC", "RFR" and "RFC".
}
\details{
SVM (support vector machine) and RF (random forest) models can be fitted in this function, including 2 classification (RFC, SVC) and 2 regression (SVR, SVC) models.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## Fit RFR model ##
machine_model <- GSmachine(markers = Markers, pheVal = phenotype, modelMethods = "RFR")

## Fit classification model(RFC) ##
machine_model <- GSmachine(markers = Markers, pheVal = phenotype, modelMethods = "RFC",
                           posPercentage = 0.4, ntree = 500)
}
}
\seealso{
\pkg{e1017}
\pkg{randomForest}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{model,}
\keyword{randomforest,}
\keyword{svm}
