% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{predictGS}
\alias{predictGS}
\title{Prediction with Trained Model from Geomic Selection Model}
\usage{
predictGS(testMat, trainModel, modelMethods = "SVC", type = "fit")
}
\arguments{
\item{testMat}{(numeric)a matrix, each row is the each testing sets or new GS data individual's SNP genotypes informations.Genotypes should be coded as {0,1,2}; 0 represent AA(homozygote), 2 represent BB(homozygote) and 1 represent AB(heterozygote); missing (NA) alleles are not allowed.}

\item{trainModel}{(model)the trained model.It's type must be similar whith modelMethods.}

\item{modelMethods}{(character)the type name of training model including "bigRR","rrBLUP","LASSO","SPLS","SVC","SVR","RFC","RFR".}

\item{type}{(character)SPLS parameter,detail see package spls.\pkg{spls}.}
}
\value{
a list:  
The prediction result of testing sets which predicted through already models
}
\description{
This function can give the prediction score of a new GS data by using already model.
}
\examples{
\dontrun{
## Load example data ##
data(GYSS)

## Fit rrBLUP model ##
rrBLUP_model <- GSReModel(markers = Markers, pheVal = phenotype, modelMethods = "rrBLUP")

## Predict 1-20 subset of all example data with already rrBLUP model ## 
res <- predictGS(testMat = Markers[1:20,], trainModel = rrBLUP_model, modelMethods = "rrBLUP")
}
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{,}
\keyword{model}
\keyword{predicttion}
