% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.r
\name{sampleClassify}
\alias{sampleClassify}
\title{Generate Positive and Negative Samples for Training}
\usage{
sampleClassify(phenotype, posPercentage = 0.4, BestIndividuals = c("top",
  "middle", "buttom"))
}
\arguments{
\item{phenotype}{(numeric)the breeding values of each individual.}

\item{posPercentage}{(numeric,[0,1])phenotype of extreme individuals which expected, default 0.4}

\item{BestIndividuals}{(character)the position of expected phenotype in whole phenotypic data set."top","buttom" or "middle",default "top".}
}
\value{
A list of row number of positive and negative samples
$posSampleIndex Index of positive samples
$negSampleIndex  Index of negative samples
}
\description{
This function can be use to generate positive and negative samples for training.The positive samples represent the excellent individuals which's breeding values we expect to obtain in your research.And the negative samples represent the lower breeding values of individuals.
}
\examples{
\dontrun{
data(GYSS)
## percentage of positive samples is 0.4 ##
sampleCly <- sampleClassify(phenotype, posPercentage = 0.4, BestIndividuals = "top")
} 
}
\author{
Chuang Ma, Qian Cheng, Zhixu Qiu
}
\keyword{,}
\keyword{negative}
\keyword{positive}
